+++ src/Producer.fs
@@ -1,11 +1,13 @@
 namespace Lmc.Kafka
 
 open System
+open Microsoft.Extensions.Logging
 open Lmc.Metrics.ServiceStatus
+open Lmc.Tracing
 
 type ProducerConfiguration = {
     Connection: ConnectionConfiguration
-    Logger: Logger option
+    Logger: ILogger option
     Checker: Checker option
     MarkAsDisabled: MarkAsDisabled option
 }
@@ -50,6 +52,8 @@ type private ProduceRuntime = {
     Topic: string
     /// Acutal partition used for producing
     Partition: int
+
+    UseTracing: bool
 }
 
 type Producer =
@@ -57,13 +61,19 @@ type Producer =
         KafkaProducer: KafkaProducer
         Topic: StreamName
         Runtime: ProduceRuntime
+        Logger: ILogger option
     }
 
+    member internal this.LogDebug(message: string): unit =
+        this.Logger |> Option.iter (fun logger -> logger.LogDebug(message))
+
     member this.Flush() =
+        this.LogDebug("Flushing producer ...")
         (this.KafkaProducer |> KafkaProducer.value).Flush()
 
     member this.Close() =
         this.Flush()
+        this.LogDebug("Closing producer ...")
         (this.KafkaProducer |> KafkaProducer.value).Dispose()
 
     interface IDisposable with
@@ -82,6 +92,7 @@ module Producer =
         ProducerBuilder(config).Build() |> KafkaProducer
 
     let private createProducer (configuration: ProducerConfiguration): Producer =
+        configuration.Logger |> Option.iter (fun logger -> logger.LogDebug("Connecting producer ..."))
         {
             KafkaProducer = configuration.Connection.BrokerList |> createKafkaProducer
             Topic = configuration.Connection.Topic
@@ -89,17 +100,19 @@ module Producer =
                 BootstrapServers = configuration.Connection.BrokerList |> BrokerList.value
                 Topic = configuration.Connection.Topic |> StreamName.value
                 Partition = ProducerConfiguration.DefaultPartition
+                UseTracing = Tracer.Check.isTracerAvailable()
             }
+            Logger = configuration.Logger
         }
 
     let private createProducerWithChecker checker (configuration: ProducerConfiguration): Producer =
         let maxRetries = checker.MaxRetries
-        let log = configuration.Logger |> Logger.resolve
         let markAsDisabled = configuration.MarkAsDisabled |> ServiceStatus.resolveMarkAsDisabled
 
         let mutable attempt = 1<Attempt>
         let mutable waitForResource = checker.WaitForResourceDefault
 
+        configuration.Logger |> Option.iter (fun logger -> logger.LogDebug("Connecting producer ..."))
         let (KafkaProducer producer) = createKafkaProducer configuration.Connection.BrokerList
 
         seq {
@@ -113,9 +126,12 @@ module Producer =
                             BootstrapServers = configuration.Connection.BrokerList |> BrokerList.value
                             Topic = configuration.Connection.Topic |> StreamName.value
                             Partition = ProducerConfiguration.DefaultPartition
+                            UseTracing = Tracer.Check.isTracerAvailable()
                         }
+                        Logger = configuration.Logger
                     }
                 | _ ->
+                    let log message = configuration.Logger |> Option.iter (fun logger -> logger.LogDebug(message))
                     let (currentAttempt, waitFor) = MarkAsDisabled.executeAndWait log attempt maxRetries markAsDisabled waitForResource
 
                     attempt <- currentAttempt
@@ -123,7 +139,9 @@ module Producer =
         }
         |> tee (fun producers ->
             if producers |> Seq.isEmpty then
-                failwithf "There is no connected producer. Problem is with either %A and/or a %A." configuration.Connection.BrokerList configuration.Connection.Topic
+                sprintf "There is no connected producer. Problem is with either %A and/or a %A." configuration.Connection.BrokerList configuration.Connection.Topic
+                |> tee (fun message -> configuration.Logger |> Option.iter (fun logger -> logger.LogError(message)))
+                |> failwith
         )
         |> Seq.head
 
@@ -163,23 +181,23 @@ module Producer =
 
     [<RequireQualifiedAccess>]
     module private Produce =
-        open Lmc.Tracing
-
         let messageWith (producer: Producer) (message: KafkaMessage) =
             let topicValue = producer.Runtime.Topic
 
             use __ =
-                "Produce event"
-                |> Trace.ChildOf.continueOrStart (Trace.extractFromKafkaHeaders message.Headers >> Trace.ofContextOption)
-                |> Trace.addTags [
-                    "peer.service", "kafka"
-                    "peer.address", producer.Runtime.BootstrapServers
-                    "component:", (sprintf "fkafka (%s)" AssemblyVersionInformation.AssemblyVersion)
-                    "kafka.topic", topicValue
-                    "message_bus.destination", topicValue
-                    "kafka.partition", string producer.Runtime.Partition
-                    "span.kind", "producer"
-                ]
+                if producer.Runtime.UseTracing then
+                    "Produce event"
+                    |> Trace.ChildOf.continueOrStart (Trace.extractFromKafkaHeaders message.Headers >> Trace.ofContextOption)
+                    |> Trace.addTags [
+                        "peer.service", "kafka"
+                        "peer.address", producer.Runtime.BootstrapServers
+                        "component:", (sprintf "fkafka (%s)" AssemblyVersionInformation.AssemblyVersion)
+                        "kafka.topic", topicValue
+                        "message_bus.destination", topicValue
+                        "kafka.partition", string producer.Runtime.Partition
+                        "span.kind", "producer"
+                    ]
+                else Inactive
 
             (producer.KafkaProducer |> KafkaProducer.value).Produce(topicValue, message)