+++ src/Types.fs
@@ -26,6 +26,7 @@ module StreamName =
         | (StreamName streamName) -> streamName
         | Instance instance -> instance |> Instance.concat "-"
 
+[<RequireQualifiedAccess>]
 type GroupId =
     | Random
     | Id of string
@@ -33,12 +34,12 @@ type GroupId =
 [<RequireQualifiedAccess>]
 module GroupId =
     let map f = function
-        | Id groupId -> groupId |> f |> Id
-        | Random -> Random
+        | GroupId.Id groupId -> groupId |> f |> GroupId.Id
+        | GroupId.Random -> GroupId.Random
 
     let value = function
-        | Id groupId -> groupId
-        | Random -> sprintf "random-%d" System.DateTime.Now.Ticks    // random (unique) group id means, it will always starts from the beginning
+        | GroupId.Id groupId -> groupId
+        | GroupId.Random -> sprintf "random-%d" System.DateTime.Now.Ticks    // random (unique) group id means, it will always starts from the beginning
 
 //
 // Headers
@@ -86,6 +87,7 @@ module Header =
 // Service Status
 //
 
+[<RequireQualifiedAccess>]
 module internal MarkAsDisabled =
     open System
 
@@ -101,6 +103,7 @@ module internal MarkAsDisabled =
 
         (currentAttempt, nextTimeWaitFor)
 
+[<RequireQualifiedAccess>]
 module internal ServiceStatus =
     let resolve = function
         | Some { MarkAsEnabled = markAsEnabled; MarkAsDisabled = markAsDisabled } -> (markAsEnabled, markAsDisabled)
@@ -109,20 +112,6 @@ module internal ServiceStatus =
     let resolveMarkAsDisabled = function
         | Some markAsDisabled -> markAsDisabled
         | _ -> MarkAsDisabled ignore
-
-//
-// Logger
-//
-
-type Logger = {
-    Log: string -> unit
-}
-
-[<RequireQualifiedAccess>]
-module internal Logger =
-    let resolve = function
-        | Some { Log = log } -> log
-        | _ -> ignore
 
 //
 // Connection
@@ -138,7 +127,7 @@ type ConnectionConfiguration = {
 //
 
 [<AutoOpen>]
-module internal GenericHelpers =
+module internal Utils =
     let tee f a =
         f a
         a